backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.14.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]>=3.3.0
